python/requirements.txt
torch
torchvision
opencv-python
numpy
Pillow
fastapi
uvicorn
gdown
python-multipart